opt: Inline Factory into Optimizer
Inline Factory into Optimizer rather than constructing it on the
heap. As part of this, collapse xfunc.CustomFuncs into norm.CustomFuncs,
since we realized we can just inline norm.CustomFuncs into the explorer
in order to get the sharing we need.

Release note: None
