+++ pkg/sql/opt/idxconstraint/index_constraints_test.go
@@ -83,7 +83,8 @@ func TestIndexConstraints(t *testing.T) {
 			var normalizeTypedExpr bool
 			var err error
 
-			f := norm.NewFactory(&evalCtx)
+			var f norm.Factory
+			f.Init(&evalCtx)
 			md := f.Metadata()
 
 			for _, arg := range d.CmdArgs {
@@ -142,7 +143,7 @@ func TestIndexConstraints(t *testing.T) {
 				for i := range varNames {
 					varNames[i] = fmt.Sprintf("@%d", i+1)
 				}
-				b := optbuilder.NewScalar(ctx, &semaCtx, &evalCtx, f)
+				b := optbuilder.NewScalar(ctx, &semaCtx, &evalCtx, &f)
 				b.AllowUnsupportedExpr = true
 				group, err := b.Build(typedExpr)
 				if err != nil {
@@ -151,7 +152,7 @@ func TestIndexConstraints(t *testing.T) {
 				ev := memo.MakeNormExprView(f.Memo(), group)
 
 				var ic idxconstraint.Instance
-				ic.Init(ev, indexCols, notNullCols, invertedIndex, &evalCtx, f)
+				ic.Init(ev, indexCols, notNullCols, invertedIndex, &evalCtx, &f)
 				result := ic.Constraint()
 				var buf bytes.Buffer
 				for i := 0; i < result.Spans.Count(); i++ {
@@ -237,7 +238,8 @@ func BenchmarkIndexConstraints(b *testing.B) {
 			if err != nil {
 				b.Fatal(err)
 			}
-			f := norm.NewFactory(nil /* evalCtx */)
+			var f norm.Factory
+			f.Init(nil /* evalCtx */)
 			md := f.Metadata()
 			for i, typ := range varTypes {
 				md.AddColumn(fmt.Sprintf("@%d", i+1), typ)
@@ -252,7 +254,7 @@ func BenchmarkIndexConstraints(b *testing.B) {
 
 			semaCtx := tree.MakeSemaContext(false /* privileged */)
 			evalCtx := tree.MakeTestingEvalContext(cluster.MakeTestingClusterSettings())
-			bld := optbuilder.NewScalar(context.Background(), &semaCtx, &evalCtx, f)
+			bld := optbuilder.NewScalar(context.Background(), &semaCtx, &evalCtx, &f)
 
 			group, err := bld.Build(typedExpr)
 			if err != nil {
@@ -262,7 +264,7 @@ func BenchmarkIndexConstraints(b *testing.B) {
 			b.ResetTimer()
 			for i := 0; i < b.N; i++ {
 				var ic idxconstraint.Instance
-				ic.Init(ev, indexCols, notNullCols, false /*isInverted */, &evalCtx, f)
+				ic.Init(ev, indexCols, notNullCols, false /*isInverted */, &evalCtx, &f)
 				_ = ic.Constraint()
 				_ = ic.RemainingFilter()
 			}
