+++ pkg/sql/opt/norm/list_builder.go
@@ -12,7 +12,7 @@
 // implied. See the License for the specific language governing
 // permissions and limitations under the License.
 
-package xfunc
+package norm
 
 import (
 	"github.com/cockroachdb/cockroach/pkg/sql/opt/memo"
